Add support for pre/post-land hooks
We now check `.git/hooks` for two executable files: `pre-land` and `post-land`.

If `pre-land` is present and executable, it is run before landing begins. If it
returns a non-zero exit code, an error is printed and `git-land` exits with that
error code. Otherwise, landing proceeds normally.

If `post-land` is present and executable, it is run after landing completes
successfully. If it returns a non-zero exit code, an error is printed and
`git-land` exits with that error code.

In each case, the hook is passed the following arguments, in order:
- the source branch
- the target branch
- the target remote

[fix #25]